Added lib files, fixed momentum Eq

+++ Source/JSph.h
@@ -178,7 +178,7 @@ protected:
   float ShiftCoef;      ///<Coefficient for shifting computation.
   float ShiftTFS;       ///<Threshold to detect free surface. Typically 1.5 for 2D and 2.75 for 3D (def=0).
 
-  double CoefH; //ABVR
+  float CoefH; //ABVR
   float Visco;  
   float ViscoBoundFactor;     //-Para interaccion con contorno usa Visco*ViscoBoundFactor.                      ///<For boundary interaction use Visco*ViscoBoundFactor.
   JSphVisco* ViscoTime;       //-Proporciona un valor de viscosidad en funcion del instante de la simulacion.   ///<Provides a viscosity value as a function of simulation time.
